requirements.txt
apache-airflow==2.10.2
beautifulsoup4==4.12.3
pandas==2.2.3
Requests==2.32.3
